+++ negotiation/src/main/java/it/eng/negotiation/service/ContractNegotiationProviderService.java
@@ -14,12 +14,14 @@ import it.eng.negotiation.model.ContractNegotiation;
 import it.eng.negotiation.model.ContractNegotiationEventMessage;
 import it.eng.negotiation.model.ContractNegotiationState;
 import it.eng.negotiation.model.ContractRequestMessage;
+import it.eng.negotiation.model.Offer;
 import it.eng.negotiation.properties.ContractNegotiationProperties;
 import it.eng.negotiation.repository.ContractNegotiationRepository;
 import it.eng.negotiation.repository.OfferRepository;
 import it.eng.negotiation.serializer.Serializer;
 import it.eng.tools.client.rest.OkHttpRestClient;
 import it.eng.tools.event.contractnegotiation.ContractNegotationOfferRequestEvent;
+import it.eng.tools.model.IConstants;
 import it.eng.tools.response.GenericApiResponse;
 import it.eng.tools.util.CredentialUtils;
 import lombok.extern.slf4j.Slf4j;
@@ -93,21 +95,32 @@ public class ContractNegotiationProviderService extends BaseProtocolService {
 			throw new OfferNotValidException("Contract offer is not valid", contractRequestMessage.getConsumerPid(), contractRequestMessage.getProviderPid());
 		}
 		
+		Offer offerToBeInserted = Offer.Builder.newInstance()
+				.assignee(contractRequestMessage.getOffer().getAssignee())
+				.assigner(contractRequestMessage.getOffer().getAssigner())
+				.originalId(contractRequestMessage.getOffer().getId())
+				.permission(contractRequestMessage.getOffer().getPermission())
+				.target(contractRequestMessage.getOffer().getTarget())
+				.build();
+		
+		Offer savedOffer = offerRepository.save(offerToBeInserted);
+		log.info("PROVIDER - Offer {} saved", savedOffer.getId());
+		
+		
         ContractNegotiation contractNegotiation = ContractNegotiation.Builder.newInstance()
                 .state(ContractNegotiationState.REQUESTED)
                 .consumerPid(contractRequestMessage.getConsumerPid())
                 .callbackAddress(contractRequestMessage.getCallbackAddress())
                 .assigner(contractRequestMessage.getOffer().getAssigner())
-                .offer(contractRequestMessage.getOffer())
+                .role(IConstants.ROLE_PROVIDER)
+                .offer(savedOffer)
                 .build();
         
         contractNegotiationRepository.save(contractNegotiation);
         log.info("PROVIDER - Contract negotiation {} saved", contractNegotiation.getId());
-
-        offerRepository.findById(contractRequestMessage.getOffer().getId())
-        	.ifPresentOrElse(o -> log.info("Offer already exists"),
-        		() -> offerRepository.save(contractRequestMessage.getOffer()));
-		log.info("PROVIDER - Offer {} saved", contractRequestMessage.getOffer().getId());
+//        offerRepository.findById(contractRequestMessage.getOffer().getId())
+//        	.ifPresentOrElse(o -> log.info("Offer already exists"),
+//        		() -> offerRepository.save(contractRequestMessage.getOffer()));
 		
 		if (properties.isAutomaticNegotiation()) {
 			log.debug("PROVIDER - Performing automatic negotiation");