+++ negotiation/src/test/java/it/eng/negotiation/service/ContractNegotiationAPIServiceTest.java
@@ -37,6 +37,7 @@ import it.eng.negotiation.repository.ContractNegotiationRepository;
 import it.eng.negotiation.repository.OfferRepository;
 import it.eng.negotiation.serializer.Serializer;
 import it.eng.tools.client.rest.OkHttpRestClient;
+import it.eng.tools.model.IConstants;
 import it.eng.tools.response.GenericApiResponse;
 import it.eng.tools.util.CredentialUtils;
 
@@ -62,6 +63,8 @@ public class ContractNegotiationAPIServiceTest {
 	private ArgumentCaptor<ContractNegotiation> argCaptorContractNegotiation;
 	@Captor
 	private ArgumentCaptor<Agreement> argCaptorAgreement;
+	@Captor
+	private ArgumentCaptor<Offer> argCaptorOffer;
 
 	@InjectMocks
 	private ContractNegotiationAPIService service;
@@ -74,11 +77,14 @@ public class ContractNegotiationAPIServiceTest {
 		when(apiResponse.getData()).thenReturn(Serializer.serializeProtocol(MockObjectUtil.CONTRACT_NEGOTIATION_ACCEPTED));
 		when(apiResponse.isSuccess()).thenReturn(true);
 		when(properties.consumerCallbackAddress()).thenReturn(MockObjectUtil.CALLBACK_ADDRESS);
+		when(offerRepository.save(any(Offer.class))).thenReturn(MockObjectUtil.OFFER_WITH_ORIGINAL_ID);
 		
 		service.startNegotiation(MockObjectUtil.FORWARD_TO, Serializer.serializePlainJsonNode(MockObjectUtil.OFFER));
 		
-		verify(contractNegotiationRepository).save(any(ContractNegotiation.class));
-		verify(offerRepository).save(any(Offer.class));
+		verify(offerRepository).save(argCaptorOffer.capture());
+		verify(contractNegotiationRepository).save(argCaptorContractNegotiation.capture());
+		assertEquals(IConstants.ROLE_CONSUMER, argCaptorContractNegotiation.getValue().getRole());
+		assertEquals(MockObjectUtil.OFFER.getId(), argCaptorOffer.getValue().getOriginalId());
 	}
 	
 	@Test
@@ -352,10 +358,12 @@ public class ContractNegotiationAPIServiceTest {
 		when(okHttpRestClient.sendRequestProtocol(any(String.class), any(JsonNode.class), any(String.class))).thenReturn(apiResponse);
 		when(apiResponse.isSuccess()).thenReturn(true);
 		
-		service.handleContractNegotiationAgreed(contractNegotaitionId);
+		service.approveContractNegotiation(contractNegotaitionId);
 		
+		verify(agreementRepository).save(argCaptorAgreement.capture());
 		verify(contractNegotiationRepository).save(argCaptorContractNegotiation.capture());
 		assertEquals(ContractNegotiationState.AGREED, argCaptorContractNegotiation.getValue().getState());
+		assertEquals(argCaptorAgreement.getValue().getId(), argCaptorContractNegotiation.getValue().getAgreement().getId());
 	}
 	
 	@Test
@@ -365,7 +373,7 @@ public class ContractNegotiationAPIServiceTest {
 		when(contractNegotiationRepository.findById(contractNegotaitionId)).thenReturn(Optional.of(MockObjectUtil.CONTRACT_NEGOTIATION_AGREED));
 		
 		assertThrows(ContractNegotiationAPIException.class, 
-				() -> service.handleContractNegotiationAgreed(contractNegotaitionId));
+				() -> service.approveContractNegotiation(contractNegotaitionId));
 	}
 	
 	@Test
@@ -380,7 +388,7 @@ public class ContractNegotiationAPIServiceTest {
 		when(apiResponse.isSuccess()).thenReturn(false);
 		
 		assertThrows(ContractNegotiationAPIException.class, 
-				() -> service.handleContractNegotiationAgreed(contractNegotaitionId));
+				() -> service.approveContractNegotiation(contractNegotaitionId));
 		
 		verify(contractNegotiationRepository, times(0)).save(argCaptorContractNegotiation.capture());
 		verify(agreementRepository, times(0)).save(argCaptorAgreement.capture());